Connector service refactor
