+++ Makefile
@@ -78,7 +78,7 @@ run-docker:
 	docker run --env "elasticsearch.hosts=http://host.docker.internal:9200" --env "elasticsearch.api_key=$(API_KEY)" --rm -it connectors
 
 exec_app:
-	cd lib/app; bundle _$(shell cat .bundler-version)_ exec ruby app.rb
+	cd lib/app; bundle _$(shell cat .bundler-version)_ exec ruby -r "./application.rb" -e "App::Application.run!"
 
 exec_cli:
 	cd lib/app; bundle _$(shell cat .bundler-version)_ exec ruby console_app.rb
