+++ bin/connectors_service
@@ -1,4 +1,4 @@
 #!/usr/bin/env ruby
 
-require 'connectors_service'
-ConnectorsService.run!
+require 'app/application'
+App::Application.run!
