+++ lib/app/application.rb
@@ -15,11 +15,15 @@ require 'utility/environment'
 require 'utility/logger'
 
 module App
-  Utility::Environment.set_execution_environment(App::Config) do
-    App::PreflightCheck.run!
-    App::Dispatcher.start!
-  rescue App::PreflightCheck::CheckFailure => e
-    Utility::Logger.error("Preflight check failed: #{e.message}")
-    exit(-1)
+  class Application
+    def self.run!
+      Utility::Environment.set_execution_environment(App::Config) do
+        App::PreflightCheck.run!
+        App::Dispatcher.start!
+      rescue App::PreflightCheck::CheckFailure => e
+        Utility::Logger.error("Preflight check failed: #{e.message}")
+        exit(-1)
+      end
+    end
   end
 end
